Release/20230919 (#180)
* png 해상도 맞추거나 svg로 변경 (#169)

* 마케팅 정보 활용 동의 약관 추가 (#170)

* 배지 dialog 간격 수정 (#171)

* 배지 dialog layout 수정

* 배지 모두보기 클릭시 dialog dismiss

* 완료 버튼 클릭시 키보드 내려가도록 하기

* Kakao 탈퇴 후 재 진입시 동의 화면 노출 (#172)

* 캘린더 기기대응 (#174)

* 요일 width 기기대응

* 캘린더 아이템 기기대응

* 일부 배지 모두보기에서 삭제 (#175)

* 기존 닉네임과 동일할시 다음 버튼 비활성화 (#176)

* 네트워크 원활하지 않을시 더미 데이터 잠깐 뜨는 버그 수정 (#177)

* intent 관련 버그 수정 (#178)

* 마이페이지 intent 수정

* 일기 상세보기 intent 수정

* 일기 작성 intent 수정

* 한국어부터 작성 step1->step2 버그 수정

* 캘린더 깜빡임 해결 (#179)

* 캘린더 깜빡임 해결

* 왼쪽으로 스와이프할때도 (다음주로 이동시) 서버통신 결과 반영

---------

+++ app/src/main/java/com/sopt/smeem/presentation/detail/DiaryDetailBottomSheet.kt
@@ -41,8 +41,8 @@ class DiaryDetailBottomSheet(
 
     fun addListeners() {
         binding.tvEdit.setOnClickListener {
-            moveToEdit()
             dismiss()
+            moveToEdit()
         }
         binding.tvDelete.setOnClickListener {
             dismiss()
@@ -51,18 +51,18 @@ class DiaryDetailBottomSheet(
     }
 
     private fun moveToEdit() {
-        val intent = Intent(requireContext(), DiaryEditActivity::class.java).apply {
+        Intent(requireContext(), DiaryEditActivity::class.java).apply {
             putExtra("diaryId", viewModel.diaryId)
             putExtra("originalContent", viewModel.diary.value)
             putExtra("randomTopic", viewModel.topic.value)
-        }
-        startActivity(intent)
+        }.run(::startActivity)
     }
 
     private fun showDeleteDialog() {
         MaterialAlertDialogBuilder(requireContext())
             .setCustomTitle(layoutInflater.inflate(R.layout.delete_dialog_title, null))
-            .setNegativeButton("예") { dialog, which ->
+            .setNegativeButton("아니요") { dialog, which -> }
+            .setPositiveButton("예") { dialog, which ->
                 viewModel.deleteDiary(
                     onSuccess = {
                         viewModel.isDiaryDeleted.value = true
@@ -73,7 +73,6 @@ class DiaryDetailBottomSheet(
                     }
                 )
             }
-            .setPositiveButton("아니요") { dialog, which -> }
             .show()
     }
